fix(create-adapter): doesn't work with mongoAdapter (#3203)
* fix(create-adapter): doesn't work with monogAdapter

In Better-auth, we do not allow customising IDs, so nothing takes into account of "different ids".
However MongoDB requires ids to be different.

The `getDefaultFieldName` fn in the create-adapter doesn't take into account that IDs from Mongo are not the same as the normal `id`. This is a patch fix until proper resolve is implemented.

* chore: lint

* chore: cleanup